project/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28
PyMySQL==1.1.0
cryptography==42.0.5